requirements.txt
annotated-types
anyio
Authlib
bandit
black
blinker
certifi
cffi
charset-normalizer
click
coverage
cryptography
Deprecated
dparse
filelock
flake8
Flask
flask-cors
Flask-Limiter
gevent
greenlet
gunicorn
h11
httpcore
httpx
idna
iniconfig
isort
itsdangerous
Jinja2
joblib
limits
markdown-it-py
MarkupSafe
marshmallow
mccabe
mdurl
mypy
mypy-extensions
nltk
ollama
ordered-set
packaging
pathspec
pbr
platformdirs
pluggy
psutil
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
pytest
pytest-cov
python-dotenv
python-json-logger
PyYAML
regex
requests
rich
ruamel.yaml
ruamel.yaml.clib
safety
safety-schemas
setuptools
shellingham
sniffio
stevedore
structlog
tqdm
typer
typing_extensions
urllib3
Werkzeug
whitenoise
wrapt
zope.event
zope.interface